[MRF-13.5] PLU-520: (frontend) refactor contexts (#1292)
## changes

1. Moving UI components into a dedicated `components` directory
2. Extracting step display logic into a new `StepsToDisplayContext`
3. Replacing the `StepExecutionsToIncludeContext` with the more comprehensive `StepsToDisplayContext`
4. Simplifying the `MrfContextProvider` by removing the explicit steps prop
5. Creating a dedicated `StepsList` component to handle the rendering of steps (Editor component was getting bloated)
